Drop apexsink support and fix include references due to newer Make.
OpenSSL usage was strictly for the apexsink support. This was for AirPort Express support, which was discontinued in 2018, and does not compile against OpenSSL 1.1.x. Given 0.10 series of gstreamer-plugins-bad has long been unsupported and will eventually be phased out / removed from the Solus repo, just go ahead and disable whatever functionality we need to just keep this building until then.